+++ csvkit/utilities/csvformat.py
@@ -22,6 +22,10 @@ class CSVFormat(CSVKitUtility):
             '-T', '--out-tabs', dest='out_tabs', action='store_true',
             help='Specify that the output file is delimited with tabs. Overrides "-D".')
         self.argparser.add_argument(
+            '-A', '--out-asv', dest='out_asv', action='store_true',
+            help='Specify that the output file is delimited with the ASCII unit separator and record separator. '
+                 'Overrides "-T", "-D" and "-M".')
+        self.argparser.add_argument(
             '-Q', '--out-quotechar', dest='out_quotechar',
             help='Character used to quote strings in the output file.')
         self.argparser.add_argument(
@@ -45,12 +49,19 @@ class CSVFormat(CSVKitUtility):
         if self.args.line_numbers:
             kwargs['line_numbers'] = True
 
-        if self.args.out_tabs:
+        if self.args.out_asv:
+            kwargs['delimiter'] = '\x1f'
+        elif self.args.out_tabs:
             kwargs['delimiter'] = '\t'
         elif self.args.out_delimiter:
             kwargs['delimiter'] = self.args.out_delimiter
 
-        for arg in ('quotechar', 'quoting', 'doublequote', 'escapechar', 'lineterminator'):
+        if self.args.out_asv:
+            kwargs['lineterminator'] = '\x1e'
+        elif self.args.out_lineterminator:
+            kwargs['lineterminator'] = self.args.out_lineterminator
+
+        for arg in ('quotechar', 'quoting', 'doublequote', 'escapechar'):
             value = getattr(self.args, f'out_{arg}')
             if value is not None:
                 kwargs[arg] = value